Move Synthetics field names to Observability Shared
