--- x-pack/plugins/exploratory_view/common/field_names/infra_logs.ts
@@ -1,10 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { DOCUMENT_FIELD_NAME } from '@kbn/lens-plugin/common';
-
-export const LOG_RATE = DOCUMENT_FIELD_NAME;
