--- x-pack/plugins/exploratory_view/common/field_names/infra_metrics.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export const SYSTEM_CPU_PERCENTAGE_FIELD = 'system.cpu.total.norm.pct';
-export const SYSTEM_MEMORY_PERCENTAGE_FIELD = 'system.memory.used.pct';
-export const DOCKER_CPU_PERCENTAGE_FIELD = 'docker.cpu.total.pct';
-export const K8S_POD_CPU_PERCENTAGE_FIELD = 'kubernetes.pod.cpu.usage.node.pct';
