--- x-pack/plugins/exploratory_view/common/index.ts
@@ -1,19 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export {
-  NETWORK_TIMINGS_FIELDS,
-  SYNTHETICS_BLOCKED_TIMINGS,
-  SYNTHETICS_CONNECT_TIMINGS,
-  SYNTHETICS_DNS_TIMINGS,
-  SYNTHETICS_RECEIVE_TIMINGS,
-  SYNTHETICS_SEND_TIMINGS,
-  SYNTHETICS_SSL_TIMINGS,
-  SYNTHETICS_STEP_DURATION,
-  SYNTHETICS_TOTAL_TIMINGS,
-  SYNTHETICS_WAIT_TIMINGS,
-} from './field_names/synthetics';
