+++ x-pack/plugins/exploratory_view/tsconfig.json
@@ -7,7 +7,8 @@
     "common/**/*",
     "public/**/*",
     "public/**/*.json",
-    "../../../typings/**/*"
+    "../../../typings/**/*",
+    "../observability_shared/common/field_names"
   ],
   "kbn_references": [
     "@kbn/core",
